backend/requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.4
daphne>=4.1
channels>=4.2
channels-redis>=4.2
celery[redis]>=5.4
redis>=5.2
psycopg2-binary>=2.9
python-decouple>=3.8
django-cors-headers>=4.6
Pillow>=11.0
whitenoise[brotli]>=6.7
gunicorn>=23.0
